+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# This Version 
+# Lib
 
 ## Features
 
@@ -10,3 +10,14 @@
 ## Bugs
 
 - Zooming only on point (0,0) not on mouse coordinate
+
+# Lib_2
+
+## Features
+
+- App, Layer, Shape, Polygon
+    - Shape supported polygon and multipolygon format
+    - Shape supported "holes" inside polygon
+
+## Bugs
+